Working on paging a little more

+++ source/bootloader/include/boot/disc.h
@@ -8,4 +8,4 @@ extern bool boot_disc_secondary_present;
 extern bool boot_disc_master_selected;
 
 bool boot_disc_init();
-bool boot_disc_load_kernel(uint64_t kernel_load_point);
+bool boot_disc_load_kernel(uint32_t kernel_load_point);
